# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Building Maintenance API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/residents/{id}": {
      "get": {
        "tags": [
          "Resident Management"
        ],
        "summary": "Get a resident by ID",
        "description": "Returns a resident by their unique identifier.",
        "operationId": "getOne_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resident found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          },
          "404": {
            "description": "Resident not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Resident Management"
        ],
        "summary": "Update a resident",
        "description": "Updates an existing resident by their ID.",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Resident update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentUpdateCommandDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Resident updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          },
          "404": {
            "description": "Resident not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Resident Management"
        ],
        "summary": "Delete a resident",
        "description": "Deletes a resident by their ID.",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resident deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resident not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/residential-complexes/{id}": {
      "get": {
        "tags": [
          "Residential Complex Management"
        ],
        "summary": "Get a residential complex by ID",
        "description": "Returns a residential complex by its unique identifier.",
        "operationId": "getOne_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the residential complex",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Residential complex found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentialComplexDto"
                }
              }
            }
          },
          "404": {
            "description": "Residential complex not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentialComplexDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Residential Complex Management"
        ],
        "summary": "Update a residential complex",
        "description": "Updates an existing residential complex by its ID.",
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the residential complex",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Residential complex update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentialComplexCommandDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Residential complex updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentialComplexDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentialComplexDto"
                }
              }
            }
          },
          "404": {
            "description": "Residential complex not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentialComplexDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Residential Complex Management"
        ],
        "summary": "Delete a residential complex",
        "description": "Deletes a residential complex by its ID.",
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the residential complex",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Residential complex deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Residential complex not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/residential-complexes": {
      "get": {
        "tags": [
          "Residential Complex Management"
        ],
        "summary": "List all residential complexes",
        "description": "Returns a list of all residential complexes.",
        "operationId": "listAll_2",
        "responses": {
          "200": {
            "description": "List of residential complexes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResidentialComplexDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Residential Complex Management"
        ],
        "summary": "Create a new residential complex",
        "description": "Creates a new residential complex with the provided details.",
        "operationId": "create_3",
        "requestBody": {
          "description": "Residential complex creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentialComplexCommandDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Residential complex created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentialComplexDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentialComplexDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResidentUpdateCommandDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string"
          },
          "unitNumber": {
            "type": "string",
            "minLength": 1
          },
          "unitBlock": {
            "type": "string",
            "minLength": 1
          },
          "residentialComplexId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "email",
          "fullName",
          "unitBlock",
          "unitNumber"
        ]
      },
      "ResidentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "unitNumber": {
            "type": "string"
          },
          "unitBlock": {
            "type": "string"
          },
          "residentialComplex": {
            "$ref": "#/components/schemas/ResidentialComplexDto"
          }
        }
      },
      "ResidentialComplexDto": {
        "type": "object",
        "description": "Summary information about a residential complex.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the residential complex."
          },
          "name": {
            "type": "string",
            "description": "Name of the residential complex."
          },
          "address": {
            "type": "string",
            "description": "Address of the residential complex."
          },
          "city": {
            "type": "string",
            "description": "City where the residential complex is located."
          },
          "postalCode": {
            "type": "string",
            "description": "Postal code of the residential complex."
          }
        }
      },
      "ResidentialComplexCommandDto": {
        "type": "object",
        "description": "DTO used to create or update a residential complex.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the residential complex.",
            "example": "Sunset Villas",
            "minLength": 1
          },
          "address": {
            "type": "string",
            "description": "Address of the residential complex.",
            "example": 123,
            "minLength": 1
          },
          "city": {
            "type": "string",
            "description": "City where the residential complex is located.",
            "example": "Springfield",
            "minLength": 1
          },
          "postalCode": {
            "type": "string",
            "description": "Postal code of the residential complex.",
            "example": 12345,
            "minLength": 1
          }
        },
        "required": [
          "address",
          "city",
          "name",
          "postalCode"
        ]
      },
...
      "DeleteResponseDto": {
        "type": "object",
        "description": "DTO returned after deleting a resource, containing a message about the operation result.",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message describing the result of the delete operation.",
            "example": "Residential complex deleted successfully."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}